+++ CHANGELOG.md
@@ -48,6 +48,7 @@ by adding an optional string as last argument to any `ParameterInput` `Get` or `
 - [[PR 163]](https://github.com/parthenon-hpc-lab/athenapk/pull/163) Add normalization by volume for relative B field divergence in history file
 
 ### Added (new features/APIs/variables/...)
+- Mixed reconstruction for GLMMHD: PLM for hydrodynamic variables and WENOZ for magnetic fields
 - Mixed reconstruction for GLMMHD: PLM for hydrodynamic variables and PPM for magnetic fields
 - [[PR 162]](https://github.com/parthenon-hpc-lab/athenapk/pull/162) Add pgen for cloud shattering setup
 - [[PR 158]](https://github.com/parthenon-hpc-lab/athenapk/pull/158) Update particle id handling (now automated `uint64`). Extend particle history lookback in turbulence pgen and include in turbulence test